requirements.txt
# Python version==3.13.3
fastmcp==2.10.6
ollama==0.5.1
asyncio==3.4.3
nest_asyncio==1.6.0
llama-index==0.12.52
llama-index-llms-ollama==0.6.2
llama-index-tools-mcp==0.2.6
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-google-vertexai==2.0.28
langchain_mcp_adapters==0.1.9
langchain-community==0.3.27
langchain-experimental==0.3.4
langgraph==0.6.3
ipykernel==6.30.1
duckdb==1.3.2
mcp-server-time==2025.8.4
neo4j==5.28.2
yfiles_jupyter_graphs==1.10.9
